fix(map): call all hooks before early returns in MapScanPanel; extract coverage detail constants; comment cleanup
